Add docs and minor fixes
