+++ tools/lint/src/main/java/com/grab/lint/Dependency.kt
@@ -1,0 +1,10 @@
+package com.grab.lint
+
+import java.io.File
+
+data class Dependency(
+    val name: String,
+    val android: Boolean,
+    val library: Boolean,
+    val partialDir: File,
+)
